Move DbImpl to a separate submodule called db_impl
